refactor: move ui auth out of opencode module
